Initial commit
I will be excited if I can actually make a tool that I actually need!!!

To give context, I am working on LensorOS and can't get any tools to make a GPT partitioned disk image to work on Windows. All tools are for physical disks and that scares me, I just want files. Luckily, I know how to write to a file, and also how GPT works, so put those together and the tool is done (with some argument parsing sprinkled in for the fun of it, right?).