More attempts at being something I'm not: good.
I suck;

it seems like fseek isn't working, or at least not properly; I'm getting different results based on when code is called even though I shouldn't due to the fseek specifying a place in the file to write to. I'm so confused at this point and I just want this all to be over, end my suffering please.