Add `null` type preset
Also rename gpt-image to createGPT

I feel like gpt-image would just get confused with AI media synthesis stuff.

createGPT is clear that GPT is the object that is being created. I do wish it was clearer that it was for creating disk **images** and not dealing with disks directly (although I'm sure you could using Unix' everything is a file mindset).